arena: close the wraparound holes in the allocation and emitter bounds
Every bound of the form `count + size > capacity` passes when the sum
wraps, so a large enough size satisfies the check it was meant to fail.
The arena had the same shape twice over: `aligned_offset + size` and
`sizeof(T) * count`.

Bound the operands rather than the results. `arena_allocate_bytes` now
checks `size` against ARENA_MAX_RESERVATION, which makes the existing
comparison against `reserved_size` exact instead of bypassable, and the
granularity round-up moves into the commit branch that needs it, so the
fast path stops loading `granularity` and pays for the new bound.
`arena_array_size` guards the one multiplication every array allocation
performs on caller-influenced data; its divisor is always a sizeof, so
it folds to a compare against an immediate and disappears entirely
wherever the count is a u32 or a literal. `parser_bump_allocate` takes
the same bound, which folds away completely at its sizeof call sites.

The emitters move to the remaining-space form `dwarf.c` and
`object_buffer_write` already used, and `ir.c` widens `ast.count * 3 + 1`
before the multiply: `count` is a u32, so the product wrapped in 32 bits
and reached the allocator as a small number no arena bound could object
to.

Costed against the self-host gate: stage 1 5.1203 G to 5.1373 G
(+0.33%), of which a fixed-workload cross-check attributes +3.5 M
(+0.068%) to compiler efficiency and the rest to the 1997 preprocessed
tokens the change adds to the tree. `ide bench` is neutral on an
interleaved A/B, the fixed point stays byte-identical, and the 19560
Release and sanitized tests pass. PERFORMANCE_AUDITS.md records the
variants that measured worse, the reservation-granularity invariant that
was faster but broke the rendering boundary tests, and the verified
sanitizer row that catches the counts which wrap before any allocator
sees them.

+++ src/buster/lib/compiler/ir/ir.c
@@ -2760,7 +2760,10 @@ BUSTER_GLOBAL_LOCAL IrLowered ir_lower_expression(IrBuilder* builder, AstExpress
     AnalysisTypedExpression* expression = ir_typed_expression_find(builder->body, ast);
     BUSTER_CHECK(expression);
     IrLowered* results = arena_allocate(builder->scratch_arena, IrLowered, ast.count);
-    IrExpressionTask* tasks = arena_allocate(builder->scratch_arena, IrExpressionTask, ast.count * 3 + 1);
+    // Widened before the multiply: `count` is a u32, so the product would wrap
+    // in 32 bits and reach the allocator as a small number the arena bound
+    // cannot object to.
+    IrExpressionTask* tasks = arena_allocate(builder->scratch_arena, IrExpressionTask, (u64)ast.count * 3 + 1);
     u32 task_count = 1;
     tasks[0] = (IrExpressionTask){
         .node_index = ast.count - 1,
